--- Tags analysis data analysis results.csv
@@ -1,5 +1,0 @@
-,Cleared/Flagged quantity,Analysis output
-0,Cleared,1027
-1,Flagged,973
-2,Percentage ratio of flagged,48.65 %
-3,Percentage ratio of Cleared,51.35 %
