--- Token analysis data analysis results.csv
@@ -1,4 +1,0 @@
-,Token details,Token count
-0,Unique tokens,7589
-1,Total tokens,59875
-2,Percentage of unique tokens,12.674739039665972 %
