+++ experiments/Tomy's-experiments/Simple ANN.py
@@ -1,0 +1,17 @@
+from tensorflow.keras.models import Sequential
+from tensorflow.keras.layers import Embedding, Conv1D, GlobalMaxPooling1D, Dense
+
+# Define CNN model
+model = Sequential([
+    Embedding(input_dim=500, output_dim=100, input_length=500), # Placeholder values used
+    Conv1D(filters=128, kernel_size=5, activation='relu'),
+    GlobalMaxPooling1D(),
+    Dense(64, activation='relu'),
+    Dense(1, activation='sigmoid')
+])
+
+# Compile model
+model.compile(optimizer='adam', loss='binary_crossentropy', metrics=['accuracy'])
+
+# Train model
+
